all prices as decimals

+++ checkout-pricing/src/main/java/com/tdd/Checkout.java
@@ -37,7 +37,7 @@ public class Checkout {
         cartHandler.add(unit);
     }
 
-    public int total() {
+    public double total() {
         TraceResult result = run();
 
         return result.ruleTrace().finalTotal();
