+++ checkout-pricing/src/main/java/com/tdd/calculation/dp/CandidateUtils.java
@@ -7,13 +7,13 @@ import java.util.List;
 
 public class CandidateUtils {
 
-    static List<String> createOptionsLabels(int i, int[] dp) {
+    static List<String> createOptionsLabels(int i, double[] dp) {
         List<String> optionsLabels = new ArrayList<>();
         optionsLabels.add("unitPrice x" + i + " = " + dp[i]);
         return optionsLabels;
     }
 
-    static List<String> createUnitPriceEntry(int i, int unitPrice) {
+    static List<String> createUnitPriceEntry(int i, double unitPrice) {
         List<String> best = new ArrayList<>();
         best.add(i + " x " + unitPrice + " = " + i*unitPrice + " kr");
         return best;
