+++ checkout-pricing/src/main/java/com/tdd/engine/application/rules/utility/Before.java
@@ -2,4 +2,4 @@ package com.tdd.engine.application.rules.utility;
 
 import com.tdd.tracing.debug.RuleTrace;
 
-public record Before(RuleTrace rt, int price) {}
+public record Before(RuleTrace rt, double price) {}
