+++ checkout-pricing/src/main/java/com/tdd/engine/utility/Evaluation.java
@@ -1,3 +1,3 @@
 package com.tdd.engine.utility;
 
-public record Evaluation(RuleContext context, int price) {}
+public record Evaluation(RuleContext context, double price) {}
