+++ checkout-pricing/src/main/java/com/tdd/rules/PricingOption.java
@@ -1,7 +1,7 @@
 package com.tdd.rules;
 
 public interface PricingOption {
-    int price();      // what does this package cost?
+    double price();      // what does this package cost?
     int quantity();   // how many items are consumed?
     int priority();   // lower = higher priority
     boolean stackable();
